New correct: use similar_text instead of levenshtein
